Add initial etude 3: arithmetic.
A the moment the program reads from file input and sorts the data into workable lists. It then calculates the 'truth table' of possible operators (+, *) depending on the length of the number. This list can then be processed in chunks the length of the number-1 to represent all permutations of multiplication/addition.